Add EICAR.antivirus_active? method
In case you want to ask directly and not really rely on the rescue
error. Maybe this is better than require bailing out of the gate? Or
also pretty good to check after the fact -- maybe the gem loaded but
then AV came in and nuked the file, so this gives more realtime status.